refactor: move error handler to global package
